% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{A function for graphing UN countries population from 1950 to 2020}
\usage{
CountryPopulation(country)
}
\arguments{
\item{country}{a string}
}
\value{
a graph of the specified country's estimated population
}
\description{
A function for graphing UN countries population from 1950 to 2020
}
\examples{
CountryPopulation('Switzerland')
CountryPopulation('New Zealand')
}
