% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{A data set of UN Countries World Populations from 1950 to 2020.}
\format{
a data frame with 235 observations of 72 variables
\describe{
  \item{Country Name}{The names of the UN countries}
  \item{1950:2020}{Population estimates for years `1950` to `2020`.}

}
}
\usage{
WorldPopulation
}
\description{
A data set of UN Countries World Populations from 1950 to 2020.
}
\keyword{datasets}
